requirements.txt
djangorestframework==3.15.1
openai==1.34.0

